Fix all 192 D3D12 int→uint compilation errors for Vortice 3.8.3
Agent-Logs-Url: https://github.com/winnerspiros/veldrid/sessions/61c71d30-3325-4ea0-a729-f70cefe2a5a2

+++ src/Veldrid/D3D12/D3D12DescriptorAllocator.cs
@@ -89,7 +89,7 @@ namespace Veldrid.D3D12
             var desc = new DescriptorHeapDescription
             {
                 Type = heapType,
-                DescriptorCount = (int)descriptorCount,
+                DescriptorCount = descriptorCount,
                 Flags = shaderVisible ? DescriptorHeapFlags.ShaderVisible : DescriptorHeapFlags.None,
                 NodeMask = 0
             };
